+++ src/api-utilts.d.ts
@@ -1,0 +1,8 @@
+//declare module 'axios' {
+//	export interface AxiosRequestConfig = AxiosRequestConfig & {
+//		'retries'?: {
+//			retryCount: number;
+//			lastRequestTime: number;
+//		};
+//	}
+//}